Embarrassing CSS padding fix
It's not a big deal, but still annoying. I can't easily wrap the logos on the
Access and About pages in a div because "rst." I can't assign a class to the
header created divs, etc, etc. It's a mess. Rather than add a class to every.
single. affected. image, I decided to use this semi-precise but wasteful method.