Add wide logo
The font and colors are lifted from the original Forrest Gump movie poster.

The font used is Old Standard TT, which is the closest free alternative I could
find to Fenice (which was the actual font used).

A highly motivated individual could tweak our text to more closely match Fenice.
I am not that motivated. ;-)