fix favicons; copy static/ correctly
This was broken with the move away from the Makefile (where we simply
used rsync).

STATIC_PATHS only searches in Pelican's PATH, which is set to content/.
Escaping the PATH using ../ does not make pelican happy, and causes it
to throw warnings everywhere and attempt overwriting its own static
files.

There is no way I can find of to copy an entire folder into the top-level
of output.

STATIC_PATHS only accepts dirs, and EXTRA_PATH_METADATA does not accept
patterns or wildcards for the key.

So this is the solution I've come up with: specifying each file by hand.

Sigh...