thanks logos: center align; increase spacing, increase max-height
This should be more accommodating of the many aspect ratios we have.
With the rules, the logos will fill their horizontal space most of the
time. The are center aligned vertically, so the differing heights feel
evenly weighted.

In the corner cases that they do not fill their horizontal space (i.e.
are capped by their vertical budget), they center horizontally. The
vertical budget is generous, but prevents squat logos from taking over
the screen (especially on portrait mobile).