dynamically insert JDON-LD content
Using the `defer` attribute, so it executes after the full page is
loaded.

This is better than shipping it statically, which bloats the page by
23KB. Inserting it dynamically allows to allow the rest of the page to
load and render first, while also leveraging caching on the metadata
file itself.

I still think modifying the DOM with javascript is a stupid way to do
this, considering the <link> tag exists. But JSON-LD does not seem to
support that.

Sigh...