hide closing quotes on blockquotes
If we don't declare closing quotes, the browser thinks they're nested.
The first one will be double quotes and all after are single.

So we give closing quotes to :after.

But I don't want to see the closing quotes.

If I apply "display: none" to :after, the browser interprets this as not
closing the quotes, and we go back to the original situation with a
bunch of single quotes.

So I will just unset the closing double quotes. If anything is ever
nested, it will behave correctly, and wrap the subquote in single
quotes. The top-level, though, will render as I wish: opening oversized
double quote, no closing double quotes. 5e291471cec53b010ceb576bd1a9654b3961f925