search: replace old tipue-search-lite with new version as submodule
The webpage was using an old version of tsl, which is now replaced with
the updated version, containing updates and bug fixxes, e.g. respecting
quotes for search terms, and searching terms case insensitive.

Now tipue-search-lite is a submodule, the old file tipuesearch_lite.js
is replaced with a link to the tipuesearch_lite.js from the submodule.